Move implementation to build_runner_core.
