+++ build_runner/lib/src/watcher/change_filter.dart
@@ -4,8 +4,6 @@
 import 'dart:async';
 
 import 'package:build/build.dart';
-// ignore: implementation_imports
-import 'package:build/src/internal.dart';
 import 'package:build_runner_core/build_runner_core.dart';
 // ignore: implementation_imports
 import 'package:build_runner_core/src/asset_graph/graph.dart';
