+++ build_runner_core/lib/src/asset/reader_writer.dart
@@ -6,14 +6,19 @@ import 'dart:convert';
 import 'dart:io';
 
 import 'package:build/build.dart';
-// ignore: implementation_imports
-import 'package:build/src/internal.dart';
 import 'package:glob/glob.dart';
 import 'package:glob/list_local_fs.dart';
 import 'package:path/path.dart' as path;
 
 import '../logging/timed_activities.dart';
 import '../package_graph/package_graph.dart';
+import '../state/asset_finder.dart';
+import '../state/asset_path_provider.dart';
+import '../state/filesystem.dart';
+import '../state/filesystem_cache.dart';
+import '../state/generated_asset_hider.dart';
+import '../state/reader_state.dart';
+import '../state/reader_writer.dart';
 import '../util/constants.dart';
 import 'writer.dart';
