+++ build_runner_core/lib/src/generate/build_step_impl.dart
@@ -12,12 +12,18 @@ import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/element/element2.dart';
 import 'package:async/async.dart';
 import 'package:build/build.dart';
-// ignore: implementation_imports
-import 'package:build/src/internal.dart';
 import 'package:crypto/crypto.dart';
 import 'package:glob/glob.dart';
 import 'package:package_config/package_config_types.dart';
 
+import '../library_cycle_graph/phased_reader.dart';
+import '../state/asset_finder.dart';
+import '../state/asset_path_provider.dart';
+import '../state/filesystem.dart';
+import '../state/filesystem_cache.dart';
+import '../state/generated_asset_hider.dart';
+import '../state/reader_state.dart';
+import '../state/reader_writer.dart';
 import 'input_tracker.dart';
 import 'run_builder.dart';
 import 'single_step_reader_writer.dart';
