+++ build_runner_core/lib/src/generate/input_tracker.dart
@@ -5,8 +5,8 @@
 import 'dart:collection';
 
 import 'package:build/build.dart';
-// ignore: implementation_imports
-import 'package:build/src/internal.dart';
+
+import '../state/filesystem.dart';
 
 /// Records inputs for a `BuildStep`.
 class InputTracker {
