+++ build_runner_core/lib/src/library_cycle_graph/library_cycle_graph.dart
@@ -2,11 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:build/build.dart' hide Builder;
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 
-import '../asset/id.dart';
 import 'library_cycle.dart';
 
 part 'library_cycle_graph.g.dart';
