+++ build_runner_core/lib/src/library_cycle_graph/phased_asset_deps.dart
@@ -2,11 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:build/build.dart' hide Builder;
 import 'package:built_collection/built_collection.dart';
 import 'package:built_value/built_value.dart';
 import 'package:built_value/serializer.dart';
 
-import '../asset/id.dart';
 import 'asset_deps.dart';
 import 'phased_value.dart';
