+++ build_runner_core/lib/src/state/filesystem.dart
