+++ build_runner_core/lib/src/state/filesystem_cache.dart
@@ -6,10 +6,9 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:build_runner_core/build_runner_core.dart';
+import 'package:build/build.dart';
 
-import '../../build.dart';
-import '../asset/id.dart';
+import '../logging/timed_activities.dart';
 import 'lru_cache.dart';
 
 /// Cache for file existence and contents.
