+++ build_runner_core/lib/src/state/lru_cache.dart
