+++ build_runner_core/lib/src/state/reader_state.dart
@@ -2,9 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:build_runner_core/build_runner_core.dart';
+import 'package:build/build.dart';
 
-import '../asset/reader.dart';
 import '../asset/writer.dart';
 import 'asset_finder.dart';
 import 'asset_path_provider.dart';
