+++ build_test/lib/src/in_memory_reader_writer.dart
@@ -6,12 +6,7 @@ import 'dart:convert';
 import 'dart:typed_data';
 
 import 'package:build/build.dart';
-// ignore: implementation_imports
-import 'package:build/src/internal.dart';
-// ignore: implementation_imports
-import 'package:build_runner_core/src/asset/reader_writer.dart';
-// ignore: implementation_imports
-import 'package:build_runner_core/src/generate/input_tracker.dart';
+import 'package:build_runner_core/build_runner_core.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as p;
 import 'package:watcher/watcher.dart';
